requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Google AI (Pinned for stability)
google-auth==2.47.0
google-api-core==2.24.1
google-cloud-aiplatform>=1.62.0
google-generativeai==0.7.0
google-genai

# HTTP client (async)
httpx>=0.27.0

# Image processing
Pillow>=10.0.0

# Utilities
python-multipart>=0.0.9
python-dotenv>=1.0.0